Use $interval instead of $timeout
Protractor waits for $timeouts to complete, so there's no way to test
what's going on before that: https://github.com/angular/protractor/blob/master/docs/timeouts.md#timeouts-from-protractor
So we're using the $interval service with a count of one. It's
totally the same from the consumer's standpoint, but makes testing
these timeouts with Protractor possible.